Clean up display
